refactoring: QuestionInputArea
컴포넌트별 파일 분리 및 propTypes 추가
rename sub component of QuestionInputArea.js
